backend/requirements.txt
fastapi>=0.115.11
uvicorn>=0.15.0
PyJWT>=2.10.1
email-validator>=2.2.0
sse-starlette>=2.2.1
markitdown>=0.0.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
websockets>=10.0
pydantic~=2.7.4
requests>=2.26.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
asyncio>=3.4.3
typing-extensions>=4.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
psutil
tavily-python>=0.5.1

